[supervisor] fix #1958: allow ide frame to reload the page
