+++ components/supervisor/frontend/src/ide-frame.ts
@@ -5,6 +5,7 @@
  */
 
 import { IDEService } from "@gitpod/gitpod-protocol/lib/ide-service";
+import { DisposableCollection } from "@gitpod/gitpod-protocol/lib/util/disposable";
 import { Emitter } from "@gitpod/gitpod-protocol/lib/util/event";
 import { SupervisorServiceClient } from "./supervisor-service-client";
 
@@ -28,11 +29,15 @@ export function load(supervisorServiceClient: SupervisorServiceClient): Promise<
         frame.className = 'gitpod-frame ide';
         frame.style.visibility = 'hidden';
 
+        let terminated = false;
         let capabilities: IDECapabilities = { service: false };
         const onDidChangeEmitter = new Emitter<void>();
         let _delegate: IDEService | undefined;
         const service: IDEService = {
             get state() {
+                if (terminated) {
+                    return 'terminated';
+                }
                 if (capabilities.service) {
                     return _delegate?.state || 'init';
                 }
@@ -45,18 +50,21 @@ export function load(supervisorServiceClient: SupervisorServiceClient): Promise<
             capabilities = capabilitiesElementAttribute && JSON.parse(capabilitiesElementAttribute) || capabilities;
             resolve({ frame, service });
         }
+        const toDisposeOnSetDelegate = new DisposableCollection();
         Object.defineProperty(window.gitpod, 'ideService', {
             get() {
                 return _delegate;
             },
-            set(delegate: IDEService) {
-                if (_delegate) {
-                    console.error(new Error('ideService is already set'));
+            set(delegate: IDEService | undefined) {
+                if (delegate === _delegate) {
                     return;
                 }
+                toDisposeOnSetDelegate.dispose();
                 _delegate = delegate;
                 onDidChangeEmitter.fire();
-                delegate.onDidChange(() => onDidChangeEmitter.fire());
+                if (delegate) {
+                    toDisposeOnSetDelegate.push(delegate.onDidChange(() => onDidChangeEmitter.fire()));
+                }
             }
         });
 
@@ -64,17 +72,30 @@ export function load(supervisorServiceClient: SupervisorServiceClient): Promise<
             console.info('IDE backend and content are ready, attaching IDE frontend...');
             document.body.appendChild(frame);
             frame.contentWindow?.addEventListener('DOMContentLoaded', () => {
-                frame.contentWindow?.history.replaceState(null, '', window.location.href);
-                if (frame.contentWindow) {
-                    frame.contentWindow.gitpod = window.gitpod;
-                }
+                const frameWindow = frame.contentWindow!;
+
+                //#region gitpod api
+                window.gitpod.ideService = undefined;
+                frameWindow.gitpod = window.gitpod;
+                //#endregion
+
+                //#region navigation
+                frameWindow.history.replaceState(null, '', window.location.href);
+                frameWindow.addEventListener('pagehide', () => {
+                    terminated = true;
+                    onDidChangeEmitter.fire();
+                });
+                //#endregion
+
+                //#region keyboard api
                 if (navigator.keyboard?.getLayoutMap && frame.contentWindow?.navigator.keyboard?.getLayoutMap) {
                     frame.contentWindow.navigator.keyboard.getLayoutMap = navigator.keyboard.getLayoutMap.bind(navigator.keyboard);
                 }
                 if (navigator.keyboard?.addEventListener && frame.contentWindow?.navigator.keyboard?.addEventListener) {
                     frame.contentWindow.navigator.keyboard.addEventListener = navigator.keyboard.addEventListener.bind(navigator.keyboard);
                 }
-            }, { once: true });
+                //#endregion
+            });
         });
     });
 }
