+++ components/supervisor/frontend/src/loading-frame.ts
@@ -10,13 +10,6 @@ import { MessageConnection, createMessageConnection } from 'vscode-jsonrpc/lib/m
 import { ConsoleLogger } from 'vscode-ws-jsonrpc';
 
 const serverOrigin = new URL(startUrl).origin;
-const relocateListener = (event: MessageEvent) => {
-    if (event.origin === serverOrigin && event.data.type == 'relocate' && event.data.url) {
-        window.removeEventListener('message', relocateListener);
-        window.location.href = event.data.url;
-    }
-};
-window.addEventListener('message', relocateListener, false);
 
 export function load(): Promise<{
     frame: HTMLIFrameElement,
